Assets/Main/Scripts/Python/requirements.txt
absl-py==0.11.0
alembic==1.4.3
altgraph==0.17.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.4.2
astunparse==1.6.3
asyncio==3.4.3
attrs==21.4.0
audioread==2.1.9
autopep8==1.5.4
backcall==0.2.0
become-yukarin @ git+https://github.com/Hiroshiba/become-yukarin@99a4998f4b7b9def2079c42be0edfc70201a1856
black==22.1.0
bleach==3.2.1
bleak==0.10.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chainer==5.4.0
chainerui==0.11.0
chardet==4.0.0
click==8.0.4
colorama==0.4.4
cupy==8.3.0
cupy-cuda102==7.8.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.0
docutils==0.16
entrypoints==0.4
fastdtw==0.3.4
fastrlock==0.5
filelock==3.0.12
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gevent==20.12.1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.32.0
h5py==2.10.0
idna==2.10
import==0.0.1
importlib-metadata==3.3.0
importlib-resources==5.4.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.7.0
itsdangerous==1.1.0
japanize-matplotlib==1.1.3
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==21.8.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
libclang==13.0.0
librosa==0.6.3
llvmlite==0.31.0
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mediapipe==0.8.9.1
mistune==0.8.4
msgpack==1.0.2
mypy==0.931
mypy-extensions==0.4.3
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numba==0.48.0
numpy==1.21.5
oauthlib==3.1.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==20.8
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pefile==2021.9.3
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.6.1
platformdirs==2.5.1
plotly==5.6.0
pooch==1.3.0
prometheus-client==0.13.1
prompt-toolkit==3.0.18
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/nc271/WorkSpace/MyApp/Python/realtime-yukarin-master/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.3
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
pylint==2.6.0
pynput==1.7.2
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyrsistent==0.18.1
pysptk==0.1.18
python-dateutil==2.8.1
python-editor==1.0.4
pythonnet==2.5.1
pytz==2021.3
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==2.0.5
pyworld==0.2.12
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.2.2
QtPy==2.0.1
readme-renderer==28.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resampy==0.2.2
rfc3986==1.4.0
rsa==4.6
scikit-learn==0.24.0
scipy==1.6.0
Send2Trash==1.8.0
six==1.15.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.22
structlog==20.2.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-chainer @ git+https://github.com/neka-nat/tensorboard-chainer@e537dc0712fb282b247986aea6105ded7b3beb59
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.8.0
tensorflow-estimator==2.4.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==2.1.0
toml==0.10.2
tomli==2.0.1
torch==1.11.0
tornado==6.1
tqdm==4.55.1
traitlets==5.0.5
twine==3.3.0
typed-ast==1.4.2
typing_extensions==4.1.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.2
world4py==0.1.1
wrapt==1.12.1
yukarin @ git+https://github.com/Hiroshiba/yukarin@85f39e21cc8f393396f8c575dec4e297d352dd5a
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
